templated map construction
The idea here is to pre-fill an array "template" with the static parts of the map (currently limited to keywords).

To create the map, you push all the dynamic parts of the map onto the stack.  The method handle executed knows which slots on the template it needs to fill in with the dynamic args.

(internally, it uses an int bitmap to mark which slots of the array template are dynamic vs static)

The same scheme could be done for vector expressions.

The bootstrap method takes the #Keyvals, a bitmap describing the positions of constant args, ... Keyword strings

The bootstrap method then links to a method handle that closes over the array template, an int[] of dynamic slots (decompressed from the bitmap)